requirements.txt
Django==3.0
django-filter
flake8==4.0.1
factory-boy==3.2.1
django-widget-tweaks 
